refactor : Change class names and split class
refactor
 - Change class name Validator -> InputValidator
 - Change class name View -> InputView
 - Split class ResultView from InputView,
 - Move methods of class InputView printCurrentCarsState, printCarInfo, printWinners, printWinner into class OutputView
