+++ src/main/java/domain/Main.java
@@ -1,14 +1,14 @@
 package domain;
 
-import view.InputOutput;
+import view.InputView;
 
 import java.util.ArrayList;
 
 public class Main {
     public static void main(String[] args) {
-        String carNames = InputOutput.inputCarsName();
+        String carNames = InputView.inputCarsName();
         ArrayList<Car> carsList = Car.asCarsList(carNames);
-        int stages = InputOutput.inputStages();
+        int stages = InputView.inputStages();
 
         Racing racing = new Racing(carsList, stages);
         racing.start();
