+++ src/main/java/view/InputView.java
@@ -1,0 +1,29 @@
+package view;
+
+import java.util.Scanner;
+
+public class InputView {
+    public static Scanner scanner = new Scanner(System.in);
+
+    public static String inputCarsName() {
+        String carsName = null;
+        boolean isValidCarsName = false;
+        while (!isValidCarsName) {
+            carsName = scanner.next();
+            isValidCarsName = InputValidator.validateCarsName(carsName);
+        }
+        return carsName;
+    }
+
+    public static int inputStages() {
+        String stages = null;
+        boolean isValidStages = false;
+        while (!isValidStages) {
+            stages = scanner.next();
+            isValidStages = InputValidator.validateStages(stages);
+        }
+        return Integer.parseInt(stages);
+    }
+
+
+}
